requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
pydantic-settings
google-generativeai
python-dotenv
pgvector                
pypdf                 
langchain-text-splitters 
python-multipart       